intermediate work on timing
